swapped feed fragment and postfragment(now renamed to feedtestfragment)
added additional firebase dependencies to allow authorization
