--- app/src/main/res/layout/fragment_post.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.recyclerview.widget.RecyclerView xmlns:android="http://schemas.android.com/apk/res/android"
-    android:id="@+id/post_grid"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-</androidx.recyclerview.widget.RecyclerView>
